rrr: convert reactor/future to module (rrr.future)
Header-only future.h (356 lines, FiberPromise<T>/FiberFuture<T>
templates over BoxEvent<T>) becomes the rrr.future module unit at
reactor/future.cpp. Sole consumer (rrr.hpp) switches from textual
include to `import rrr.future;`.
